Final PathScale touchups & Jacquard info Bump optimization level to -O3 Massage gasneti_atomic_compare_and_swap to work under PathScale Workaround the mysterious testsmall crash on PathScale - apparently their optimizer does something bad with default argument promotion of floats. Fix a misc harmless warning.
Note: to configure for Infiniband on Jacquard, you need to set:
MTHOME=/usr/local/ibgd/driver/infinihost/
you also need to load an mpi module compatible with the C compiler you're using.

I recommend this setup if you want gcc-based compilation:

module load pgi
module load compiler/pathscale2
# compiler modules rudely set these vars, breaking other things
unsetenv MPICH_CC
unsetenv MPICH_F77
unsetenv MPICH_CLINKER
unsetenv CC
unsetenv FC
unsetenv F77
unsetenv F90
unsetenv CXX
module unload mpi mvapich
module load mvapich/gcc

For PathScale based building, change the last line to mvapich/pathscale
CVSMAIL: CC:

+++ tests/testlarge.c
@@ -1,6 +1,6 @@
 /*   $Source: /Users/kamil/work/gasnet-cvs2/gasnet/tests/testlarge.c,v $
- *     $Date: 2005/03/12 11:21:16 $
- * $Revision: 1.28 $
+ *     $Date: 2005/03/12 13:03:19 $
+ * $Revision: 1.29 $
  * Description: GASNet bulk get/put performance test
  *   measures the ping-pong average round-trip time and
  *   average flood throughput of GASNet bulk gets and puts
@@ -87,7 +87,7 @@ void _print_stat(int myproc, stat_struct_t *st, const char *name, int operation)
 		printf("Proc %3i - %10i byte : %7i iters,"
 			   " latency %12i us total, %9.3f us ave. (%s)\n",
 			myproc, st->datasize, st->iters, (int) st->time,
-			((float)st->time) / st->iters,
+			((double)st->time) / st->iters,
 			name);
 		fflush(stdout);
 		break;
